wrappers/a2a/requirements.txt
fastapi
uvicorn[standard]
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-fastapi
google-cloud-trace
google-cloud-logging
